pyproject.toml
[project]
name = "item-data-agent"
version = "0.1.0"
description = "AI agent for supplier communication and item data management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "imap-tools>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]
